+++ CHANGELOG.md
@@ -1,4 +1,8 @@
 # Changelog
 
+## [R240807] - 2024-08-07
+- Updated to work with new payloads for endpoints.
+- Added check job status time interval for API tiers.
+
 ## [R247275-alpha] - 2024-07-27
 - First alpha release.